Moved cache engine from dl_core to new dl_cache_engine package
