+++ app/dl_control_api/pyproject.toml
@@ -24,6 +24,10 @@ datalens-core = {path = "../../lib/dl_core"}
 datalens-app-tools = {path = "../../lib/dl_app_tools"}
 datalens-model-tools = {path = "../../lib/dl_model_tools"}
 datalens-api-lib = {path = "../../lib/dl_api_lib"}
+datalens-api-commons = {path = "../../lib/dl_api_commons"}
+datalens-cache-engine = {path = "../../lib/dl_cache_engine"}
+datalens-configs = {path = "../../lib/dl_configs"}
+datalens-constants = {path = "../../lib/dl_constants"}
 
 [build-system]
 requires = ["poetry-core"]
