+++ lib/dl_cache_engine/README.md
@@ -1,0 +1,3 @@
+# dl_cache_engine
+
+This package host the main low-level logic for DataLens backend caches.
