+++ lib/dl_cache_engine/pyproject.toml
@@ -1,27 +1,26 @@
 
 [tool.poetry]
-name = "datalens-model-tools"
+name = "datalens-cache-engine"
 version = "0.0.1"
 description = ""
 authors = ["DataLens Team <datalens-opensource@yandex-team.ru>"]
-packages = [{include = "dl_model_tools"}]
+packages = [{include = "dl_cache_engine"}]
 license = "Apache 2.0"
 readme = "README.md"
 
 
 [tool.poetry.dependencies]
 attrs = ">=22.2.0"
-datalens-constants = {path = "../dl_constants"}
-dynamic-enum = {path = "../dynamic_enum"}
-marshmallow = ">=3.19.0"
 python = ">=3.10, <3.12"
+datalens-app-tools = {path = "../dl_app_tools"}
+datalens-constants = {path = "../dl_constants"}
+datalens-model-tools = {path = "../dl_model_tools"}
+lz4 = "==4.3.2"
+redis = "==4.5.4"
+redis_cache_lock = {path = "../redis-cache-lock"}
 
 [tool.poetry.group.tests.dependencies]
 pytest = ">=7.2.2"
-
-[[tool.mypy.overrides]]
-module = "dl_model_tools.contrib.enum_field"
-ignore_errors = true
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
@@ -34,10 +33,8 @@ minversion = "6.0"
 addopts = "-ra"
 testpaths = []
 
-[datalens.pytest.unit]
-root_dir = "dl_model_tools_tests/"
-target_path = "unit"
-skip_compose = "true"
+[datalens_ci]
+skip_test = true
 
 [tool.mypy]
 warn_unused_configs = true
