+++ lib/dl_compeng_pg/dl_compeng_pg/compeng_asyncpg/exec_adapter_asyncpg.py
@@ -19,10 +19,10 @@ import attr
 import sqlalchemy as sa
 from sqlalchemy.dialects.postgresql import pypostgresql
 
+from dl_cache_engine.primitives import LocalKeyRepresentation
 from dl_compeng_pg.compeng_pg_base.exec_adapter_base import PostgreSQLExecAdapterAsync
 from dl_constants.enums import UserDataType
 from dl_core.connectors.base.error_transformer import DbErrorTransformer
-from dl_core.data_processing.cache.primitives import LocalKeyRepresentation
 from dl_core.data_processing.prepared_components.primitives import PreparedFromInfo
 from dl_core.data_processing.processing.context import OpExecutionContext
 from dl_core.data_processing.streaming import (
