+++ lib/dl_compeng_pg/pyproject.toml
@@ -14,9 +14,9 @@ aiopg = ">=1.4.0"
 asyncpg = ">=0.27.0"
 attrs = ">=22.2.0"
 python = ">=3.10, <3.12"
-shortuuid = ">=1.0.11"
 sqlalchemy = ">=1.4.46, <2.0"
 datalens-sqlalchemy-postgres = {path = "../dl_sqlalchemy_postgres"}
+datalens-cache-engine = {path = "../dl_cache_engine"}
 datalens-constants = {path = "../dl_constants"}
 datalens-connector-postgresql = {path = "../dl_connector_postgresql"}
 datalens-core = {path = "../dl_core"}
