+++ lib/dl_core/dl_core/data_processing/prepared_components/primitives.py
@@ -11,12 +11,12 @@ from typing import (
 
 import attr
 
+from dl_cache_engine.primitives import LocalKeyRepresentation
 from dl_constants.enums import (
     JoinType,
     UserDataType,
 )
 from dl_core.connectors.base.query_compiler import QueryCompiler
-from dl_core.data_processing.cache.primitives import LocalKeyRepresentation
 from dl_core.query.bi_query import SqlSourceType
