+++ lib/dl_core/dl_core/data_processing/source_builder.py
@@ -20,10 +20,10 @@ import attr
 import sqlalchemy as sa
 from sqlalchemy.sql.elements import ClauseElement
 
+from dl_cache_engine.primitives import LocalKeyRepresentation
 from dl_constants.enums import JoinType
 from dl_core import exc
 from dl_core.components.ids import AvatarId
-from dl_core.data_processing.cache.primitives import LocalKeyRepresentation
 from dl_core.data_processing.prepared_components.primitives import (
     PreparedMultiFromInfo,
     PreparedSingleFromInfo,
