+++ lib/dl_model_tools/dl_model_tools/serialization.py
