+++ metapkg/pyproject.toml
@@ -151,6 +151,7 @@ datalens-connector-ydb = {path = "../lib/dl_connector_ydb"}
 datalens-dashsql = {path = "../lib/dl_dashsql"}
 datalens-pivot = {path = "../lib/dl_pivot"}
 datalens-pivot-pandas = {path = "../lib/dl_pivot_pandas"}
+datalens-cache-engine = {path = "../lib/dl_cache_engine"}
 
 [tool.poetry.group.dev.dependencies]
 black = "==23.3.0"
